Moved controllers to a common namespace
