+++ test/FubarDev.WebDavServer.Tests/Support/Controllers/SimpleWebDavController.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="TestWebDavController.cs" company="Fubar Development Junker">
+﻿// <copyright file="SimpleWebDavController.cs" company="Fubar Development Junker">
 // Copyright (c) Fubar Development Junker. All rights reserved.
 // </copyright>
 
@@ -7,13 +7,13 @@ using FubarDev.WebDavServer.AspNetCore;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace FubarDev.WebDavServer.Tests.Support
+namespace FubarDev.WebDavServer.Tests.Support.Controllers
 {
     [Route("{*path}")]
     /* [Authorize] */
-    public class TestWebDavController : WebDavControllerBase
+    public class SimpleWebDavController : WebDavControllerBase
     {
-        public TestWebDavController(
+        public SimpleWebDavController(
             IWebDavContext context,
             IWebDavDispatcher dispatcher,
             ILoggerFactory? loggerFactory = null)
