+++ test/FubarDev.WebDavServer.Tests/Support/Controllers/WinCompatibleWebDavController.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="TestWebDavController.cs" company="Fubar Development Junker">
+﻿// <copyright file="WinCompatibleWebDavController.cs" company="Fubar Development Junker">
 // Copyright (c) Fubar Development Junker. All rights reserved.
 // </copyright>
 
@@ -7,21 +7,21 @@ using FubarDev.WebDavServer.AspNetCore;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace FubarDev.WebDavServer.Tests.Issues.Issue53
+namespace FubarDev.WebDavServer.Tests.Support.Controllers
 {
     /// <summary>
     /// The WebDAV controller.
     /// </summary>
     [Route("_dav/{*path}")]
-    public class TestWebDavController : WebDavControllerBase
+    public class WinCompatibleWebDavController : WebDavControllerBase
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="TestWebDavController"/> class.
+        /// Initializes a new instance of the <see cref="WinCompatibleWebDavController"/> class.
         /// </summary>
         /// <param name="context">The WebDAV context.</param>
         /// <param name="dispatcher">The WebDAV method dispatcher.</param>
         /// <param name="loggerFactory">The logger for the response messages.</param>
-        public TestWebDavController(
+        public WinCompatibleWebDavController(
             IWebDavContext context,
             IWebDavDispatcher dispatcher,
             ILoggerFactory? loggerFactory = null)
